% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryad_select_url.R
\name{dryad_list_csv}
\alias{dryad_list_csv}
\alias{dryad_select_url}
\title{Read files stored in dryad or find their download URLs.}
\usage{
dryad_list_csv(doi, regexp = "[.]csv$", invert = FALSE,
  ignore.case = TRUE)

dryad_select_url(dryad_url, regexp = NULL, invert = FALSE,
  ignore.case = TRUE)
}
\arguments{
\item{doi}{A character string giving a valid dryad doi, such as
"'10.5061/dryad.6nc8c'".}

\item{regexp}{A regular expression (e.g. \code{[.]csv$}) passed on to \code{\link[=grep]{grep()}} to filter paths.}

\item{invert}{If \code{TRUE} return files which do \emph{not} match}

\item{ignore.case}{if \code{FALSE}, the pattern matching is \emph{case
      sensitive} and if \code{TRUE}, case is ignored during matching.}

\item{dryad_url}{A character string. The output of \code{rdryad::dryad_files()}.}
}
\value{
When run interactively, both functions ask you to select files.
\itemize{
\item \code{dryad_list_csv()} outputs a list with the selected files.
\item \code{dryad_select_url()} outputs a character string giving the download
URLs of the selected files.
}
}
\description{
Read files stored in dryad or find their download URLs.
}
\examples{
scbi_functional_ecology_2015 <- '10.5061/dryad.6nc8c'

# Help find a relevant url
dryad_select_url(
  rdryad::dryad_files(scbi_functional_ecology_2015)
)

# Shortcut to download and read multiple .csv files at once
dryad_list_csv(scbi_functional_ecology_2015)
}
