% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scbi_elevation.R
\docType{data}
\name{scbi_elevation}
\alias{scbi_elevation}
\title{Elevation data from SCBI.}
\format{An object of class \code{fgeo_elevation} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 693 rows and 3 columns.}
\usage{
scbi_elevation
}
\description{
Elevation data from SCBI.
}
\examples{
str(scbi_elevation)

head(scbi_elevation)
}
\keyword{datasets}
